add possibility for user to edit its personnal informations and destroy its account (and all its blogs)

+++ app/workers/blog_folder_eraser_worker.rb
@@ -2,6 +2,8 @@ class BlogFolderEraserWorker
   include Sidekiq::Worker
 
   def perform(path)
-    FileUtils.rm_rf path
+    if File.exist? path
+      FileUtils.rm_rf path
+    end
   end
 end
